Use versioning based on build date and git commit
Currently, the version number for the appliance is based on
the version of libguestfs and kernel included in it.

This is informative, but becomes problematic when you consider
that everything except those two components could change in
between two builds and the publishing process would believe
that there's no need to make a fresh appliance available. This
could potentially lead to the "latest" appliance being several
months out of date and missing out on important fixes.

Switch to a versioning scheme that matches the one used for
KubeVirt's builder image.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>