Include full filename in latest-version file
Right now we're including the prefix and the version number
but leaving out the extension. Depending on how you look at
it, that's either too much or not enough information :)

Include the full filename: this will allow us to do things
like changing the compression used for tarballs in a manner
that is mostly transparent to consumers.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>